+++ src/view/trip-point-edit.js
@@ -1,7 +1,7 @@
 import he from 'he';
 import flatpickr from 'flatpickr';
 import {POINT_BLANK} from '../const.js';
-import {generateOffers, generateDestination, generateTypePoint, generateCityPoint, getDuration, getDurationTripPoint} from '../utils/common.js';
+import {getDuration, getDurationTripPoint} from '../utils/common.js';
 import {replace} from '../utils/render.js';
 import SmartView from './smart.js';
 import TripDestinationsModel from '../model/destinations.js';
@@ -138,12 +138,12 @@ export default class TripPointEdit extends SmartView {
     this._point = TripPointEdit.tripPointToData(point);
 
     if (point === POINT_BLANK) {
-      const cityPoint = generateCityPoint();
-      const typePoint = generateTypePoint();
-      const pointOffers = generateOffers(typePoint);
-      const destination = generateDestination(cityPoint);
+      const cityPoint = TripDestinationsModel.getRandomCityPoint();
+      const typePoint = TripOffersModel.getRandomTypePoint();
+      const pointOffers = TripOffersModel.getTypeOffers(typePoint);
+      const destination = TripDestinationsModel.getCityDestination(cityPoint);
       const startDateTime = '2021-09-17T00:44:07.377Z';
-      const endDateTime = '2021-09-17T06:16:46.875Z';
+      const endDateTime = '2021-09-18T00:44:07.377Z';
       const duration = getDurationTripPoint(1, 0, 0);
 
       this.updateData({
@@ -222,8 +222,8 @@ export default class TripPointEdit extends SmartView {
         {},
         data,
         {
-          pointOffers: generateOffers(data.typePoint.toLowerCase()),
-          destination: generateDestination(data.cityPoint),
+          pointOffers: TripOffersModel.getTypeOffers(data.typePoint.toLowerCase()),
+          destination: TripDestinationsModel.getCityDestination(data.cityPoint),
         },
       );
     } else if (data.isChangeTripPointType) {
@@ -233,7 +233,7 @@ export default class TripPointEdit extends SmartView {
         {},
         data,
         {
-          pointOffers: generateOffers(data.typePoint.toLowerCase()),
+          pointOffers: TripOffersModel.getTypeOffers(data.typePoint.toLowerCase()),
         },
       );
     } else if (data.isChangeTripPointCity) {
@@ -243,7 +243,7 @@ export default class TripPointEdit extends SmartView {
         {},
         data,
         {
-          destination: generateDestination(data.cityPoint),
+          destination: TripDestinationsModel.getCityDestination(data.cityPoint),
         },
       );
     }
